Fix arrow key bug
