+++ src/components/UrlUpload/index.tsx
@@ -6,17 +6,14 @@ import { TUTORIAL_PATHNAME, VIEWER_PATHNAME } from "../../routes";
 
 const styles = require("./style.css");
 
-const UrlUpload = (): JSX.Element => {
-    const [isModalVisible, setIsModalVisible] = useState(false);
+interface UrlUploadProps {
+    setIsModalVisible: (isModalVisible: boolean) => void;
+}
+
+const UrlUpload = ({ setIsModalVisible }: UrlUploadProps): JSX.Element => {
     const [userInput, setUserInput] = useState("");
     const history = useHistory();
 
-    const showModal = () => {
-        if (history.location.pathname !== VIEWER_PATHNAME) {
-            history.push(VIEWER_PATHNAME);
-        }
-        setIsModalVisible(true);
-    };
     const closeModal = () => {
         setIsModalVisible(false);
     };
@@ -45,59 +42,50 @@ const UrlUpload = (): JSX.Element => {
     };
 
     return (
-        <>
-            <Button type="ghost" onClick={showModal}>
-                From a URL
-            </Button>
-            {/* Using isModalVisible this way instead of as a `visible` prop forces
-            Modal to re-render every time it is opened, resetting the form */}
-            {isModalVisible && (
-                <Modal
-                    className={styles.modal}
-                    title="Load Model from URL"
-                    visible
-                    footer={null}
-                    onCancel={closeModal}
-                    width={525}
-                    centered
+        <Modal
+            className={styles.modal}
+            title="Load Model from URL"
+            visible
+            footer={null}
+            onCancel={closeModal}
+            width={525}
+            centered
+        >
+            <Form
+                layout="vertical"
+                requiredMark={false}
+                onFinish={loadTrajectory}
+            >
+                <Form.Item
+                    name="url"
+                    label="Enter the URL to a public .simularium file"
+                    extra={extraInfo}
+                    rules={[
+                        {
+                            type: "url",
+                            message: "!\u20DD Please input a valid URL",
+                        },
+                    ]}
                 >
-                    <Form
-                        layout="vertical"
-                        requiredMark={false}
-                        onFinish={loadTrajectory}
+                    <Input
+                        allowClear
+                        placeholder="https://.../example.simularium"
+                        size="large"
+                        onChange={handleUserInput}
+                        // autofocus FIXME: this doesn't work
+                    />
+                </Form.Item>
+                <Form.Item className={styles.submitButton}>
+                    <Button
+                        type="default"
+                        htmlType="submit"
+                        disabled={!userInput}
                     >
-                        <Form.Item
-                            name="url"
-                            label="Enter the URL to a public .simularium file"
-                            extra={extraInfo}
-                            rules={[
-                                {
-                                    type: "url",
-                                    message: "!\u20DD Please input a valid URL",
-                                },
-                            ]}
-                        >
-                            <Input
-                                allowClear
-                                placeholder="https://.../example.simularium"
-                                size="large"
-                                onChange={handleUserInput}
-                                autoFocus
-                            />
-                        </Form.Item>
-                        <Form.Item className={styles.submitButton}>
-                            <Button
-                                type="default"
-                                htmlType="submit"
-                                disabled={!userInput}
-                            >
-                                Load
-                            </Button>
-                        </Form.Item>
-                    </Form>
-                </Modal>
-            )}
-        </>
+                        Load
+                    </Button>
+                </Form.Item>
+            </Form>
+        </Modal>
     );
 };
